requirements.txt
decorator==4.1.2
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
graphene==2.0.1
graphene-django==2.0.0
graphene-django-extras==0.1.2
graphql-core==2.0
graphql-relay==0.4.5
ipython==6.2.1
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.11.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pkg-resources==0.0.0
promise==2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
Rx==1.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
